requirements.txt
# BrowserControl Requirements
# Python 3.13.8
# All packages have full Python 3.13 support with pre-built wheels

# Browser Automation
selenium==4.26.1
webdriver-manager==4.0.2

# HTTP & API
requests==2.32.3
Flask==3.0.3

# Authentication
ldap3==2.9.1
pycryptodome==3.21.0

# Database
pyodbc==5.2.0
SQLAlchemy==2.0.36

# UI & System
pystray==0.19.5
Pillow==11.0.0
pywin32==308
pygetwindow==0.0.9
pyautogui==0.9.54

# Build Tools (for creating executables)
# Install separately: pip install pyinstaller==6.11.0
# pyinstaller==6.11.0
